server/requirements.txt
flask==1.1.2
flask-graphql==2.0.1
flask-marshmallow==0.14.0
flask-testing==0.8.1
flask-sqlalchemy==2.5.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
graphene==2.1.9
flask-cors==3.0.10
pymysql==1.0.2
pyjwt==2.1.0
sqlalchemy==1.4.23
pylint==2.10.2
